Crate servo_constellation

Modules

broadcastchannel 🔒

browsingcontext 🔒

constellation 🔒

The Constellation, Servo’s Grand Central Station

constellation_webview 🔒

embedder 🔒

event_loop 🔒

This module contains the EventLoop type, which is the constellation’s view of a script thread. When an EventLoop is dropped, an ExitScriptThread message is sent to the script thread, asking it to shut down.

logging 🔒

The constellation uses logging to perform crash reporting. The constellation receives all warn!, error! and panic! messages, and generates a crash report when it receives a panic.

pipeline 🔒

process_manager 🔒

sandboxing 🔒

screenshot_readiness_request 🔒

serviceworker 🔒

session_history 🔒

tracing 🔒

Structs

Constellation

The Constellation itself. In the servo browser, there is one constellation, which maintains all of the browser global data. In embedded applications, there may be more than one constellation, which are independent of each other.

EventLoop

https://html.spec.whatwg.org/multipage/#event-loop

FromEmbedderLogger

A logger directed at the constellation from Paint.

FromScriptLogger

A logger directed at the constellation from content processes #[derive(Clone)]

InitialConstellationState

State needed to construct a constellation.

NewScriptEventLoopProcessInfo

All of the information necessary to create a new script EventLoop in a new process.

Enums

ConstellationToEmbedderMsg

Messages sent from the Constellation to the embedder.

UnprivilegedContent

Functions

content_process_sandbox_profile

Our content process sandbox profile on Linux. As restrictive as possible.