servo_constellation

Module event_loop

This module contains the EventLoop type, which is the constellation’s view of a script thread. When an EventLoop is dropped, an ExitScriptThread message is sent to the script thread, asking it to shut down.

Structs

EventLoop

https://html.spec.whatwg.org/multipage/#event-loop

NewScriptEventLoopProcessInfo

All of the information necessary to create a new script EventLoop in a new process.