servo_constellation

Module logging

The constellation uses logging to perform crash reporting. The constellation receives all warn!, error! and panic! messages, and generates a crash report when it receives a panic.

Structs

FromEmbedderLogger

A logger directed at the constellation from Paint.

FromScriptLogger

A logger directed at the constellation from content processes #[derive(Clone)]

Functions

log_entry 🔒

Rust uses Record for storing logging, but servo converts that to a LogEntry. We do this so that we can record panics as well as log messages, and because Record does not implement serde (de)serialization, so cannot be used over an IPC channel.