servo_constellation::pipeline

Struct Pipeline

pub struct Pipeline {

    pub id: PipelineId,
    pub browsing_context_id: BrowsingContextId,
    pub webview_id: WebViewId,
    pub opener: Option<BrowsingContextId>,
    pub event_loop: Rc<EventLoop>,
    pub paint_proxy: PaintProxy,
    pub url: ServoUrl,
    pub animation_state: AnimationState,
    pub document_callbacks_active: bool,
    pub worker_callbacks_active: FxHashSet<WorkerId>,
    pub last_callbacks_active_sent_to_paint: bool,
    pub children: Vec<BrowsingContextId>,
    pub load_data: LoadData,
    pub history_state_id: Option<HistoryStateId>,
    pub history_states: HashSet<HistoryStateId>,
    pub completely_loaded: bool,
    pub title: String,
    pub focus_sequence: FocusSequenceNumber,

}

A Pipeline is the constellation’s view of a Window. Each pipeline has an event loop (executed by a script thread). A script thread may be responsible for many pipelines.

Fields
id: PipelineId

The ID of the pipeline.

browsing_context_id: BrowsingContextId

The ID of the browsing context that contains this Pipeline.

webview_id: WebViewId

The WebViewId of the WebView that contains this Pipeline.

opener: Option<BrowsingContextId>event_loop: Rc<EventLoop>

The event loop handling this pipeline.

paint_proxy: PaintProxy

A channel to Paint.

url: ServoUrl

The most recently loaded URL in this pipeline. Note that this URL can change, for example if the page navigates to a hash URL.

animation_state: AnimationState

Whether this pipeline is currently running animations. Pipelines that are running animations cause composites to be continually scheduled.

document_callbacks_active: bool

Whether the document for this pipeline has active animation frame callbacks.

worker_callbacks_active: FxHashSet<WorkerId>

Workers in this pipeline with active animation frame callbacks.

last_callbacks_active_sent_to_paint: bool

The last aggregate animation-callback state sent to Paint.

children: Vec<BrowsingContextId>

The child browsing contexts of this pipeline (these are iframes in the document).

load_data: LoadData

The Load Data used to create this pipeline.

history_state_id: Option<HistoryStateId>

The active history state for this pipeline.

history_states: HashSet<HistoryStateId>

The history states owned by this pipeline.

completely_loaded: bool

Has this pipeline received a notification that it is completely loaded?

title: String

The title of this pipeline’s document.

focus_sequence: FocusSequenceNumber
Implementations

impl Pipeline

pub(crate) fn spawn<STF: ScriptThreadFactory, SWF: ServiceWorkerManagerFactory>( new_pipeline_info: NewPipelineInfo, event_loop: Rc<EventLoop>, constellation: &Constellation<STF, SWF>, throttled: bool, ) -> Result<Self, SendError>

Possibly starts a script thread, in a new process if requested.

pub fn new_already_spawned( id: PipelineId, browsing_context_id: BrowsingContextId, webview_id: WebViewId, opener: Option<BrowsingContextId>, event_loop: Rc<EventLoop>, paint_proxy: PaintProxy, throttled: bool, load_data: LoadData, ) -> Self

Creates a new Pipeline, after it has been spawned in its EventLoop.

pub fn send_exit_message_to_script(&self, discard_bc: DiscardBrowsingContext)

Let the ScriptThread for this Pipeline know that it has exited. If the ScriptThread hasn’t panicked and is still alive, it will send a PipelineExited message back to the Constellation when it finishes cleaning up.

pub fn set_activity(&self, activity: DocumentActivity)

Notify this pipeline of its activity.

pub fn to_sendable(&self) -> CompositionPipeline

Paint’s view of a pipeline.

pub fn add_child(&mut self, browsing_context_id: BrowsingContextId)

Add a new child browsing context.

pub fn remove_child(&mut self, browsing_context_id: BrowsingContextId)

Remove a child browsing context.

pub fn set_throttled(&self, throttled: bool)

Set whether to make pipeline use less resources, by stopping animations and running timers at a heavily limited rate.