servo_constellation

Struct EventLoop

pub struct EventLoop {
    script_chan: GenericSender<ScriptThreadMessage>,
    id: ScriptEventLoopId,
    background_hang_monitor_sender: Option<GenericSender<BackgroundHangMonitorControlMsg>>,
    dont_send_or_sync: PhantomData<Rc<()>>,
}

https://html.spec.whatwg.org/multipage/#event-loop

Fields
script_chan: GenericSender<ScriptThreadMessage>id: ScriptEventLoopIdbackground_hang_monitor_sender: Option<GenericSender<BackgroundHangMonitorControlMsg>>

When running in another process, this is an IpcSender to the BackgroundHangMonitor on the other side of the process boundary. When running in the same process, the BackgroundHangMonitor is shared among all EventLoops so this will be None.

dont_send_or_sync: PhantomData<Rc<()>>
Implementations

impl EventLoop

pub(crate) fn spawn<STF: ScriptThreadFactory, SWF: ServiceWorkerManagerFactory>( constellation: &mut Constellation<STF, SWF>, is_private: bool, ) -> Result<Rc<Self>, IpcError>

fn spawn_in_thread<STF: ScriptThreadFactory, SWF: ServiceWorkerManagerFactory>( constellation: &mut Constellation<STF, SWF>, initial_script_state: InitialScriptState, ) -> Self

fn spawn_in_process<STF: ScriptThreadFactory, SWF: ServiceWorkerManagerFactory>( constellation: &mut Constellation<STF, SWF>, initial_script_state: InitialScriptState, ) -> Result<Self, IpcError>

pub(crate) fn id(&self) -> ScriptEventLoopId

pub fn send(&self, msg: ScriptThreadMessage) -> Result<(), SendError>

Send a message to the event loop.

pub(crate) fn send_message_to_background_hang_monitor( &self, message: &BackgroundHangMonitorControlMsg, )

If this is EventLoop is in another process, send a message to its BackgroundHangMonitor, otherwise do nothing.